requirements.txt
asn1crypto==0.24.0
astroid==2.4.2
attrs==17.4.0
Automat==0.6.0
blinker==1.4
certifi==2020.6.20
chardet==3.0.4
click==6.7
cloud-init
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
configparser==5.0.1
constantly==15.1.0
crayons==0.4.0
cryptography==2.1.4
distro-info===0.18ubuntu0.18.04.1
ec2-hibinit-agent==1.0.0
hibagent==1.0.1
httplib2==0.9.2
hyperlink==17.3.1
idna==2.10
incremental==16.10.1
isort==5.7.0
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
lazy-object-proxy==1.4.3
lxml==4.6.1
MarkupSafe==1.0
mccabe==0.6.1
netifaces==0.10.4
numpy==1.19.4
oauthlib==2.0.6
PAM==0.4.2
pandas==1.1.4
pika==1.2.0
Pillow==8.0.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
pygobject==3.26.1
PyJWT==1.5.3
pylint==2.6.0
pyOpenSSL==17.5.0
PyPDF2==1.26.0
pyserial==3.4
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2020.4
pyxdg==0.25
PyYAML==3.12
requests==2.24.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
selenium==3.141.0
service-identity==16.0.0
six==1.15.0
ssh-import-id==5.7
systemd-python==234
toml==0.10.2
Twisted==17.9.0
typed-ast==1.4.2
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.11
virtualenv==15.1.0
webdriver-manager==3.2.2
wrapt==1.12.1
zope.interface==4.3.2